fix : PostService 매개변수 타입 변경으로 인한 수정

+++ src/main/java/com/codessquad/qna/controller/PostController.java
@@ -57,7 +57,7 @@ public class PostController {
      * @return
      */
     @GetMapping("/questions/{id}")
-    public String getPost(@PathVariable int id, Model model) {
+    public String getPost(@PathVariable Long id, Model model) {
         try{
             Post post = postService.getPost(id);
             model.addAttribute("post", post);
